Fitting Data for Interval Calculation

This function takes data from the prep_interval_data function and fits it to a series of speficied distributions

fit_interval_data(
  interval_data,
  distribution = c("gamma", "gamma-trunc"),
  stan_opts = list(),
  return_raw = FALSE
)

Arguments

interval_data
a named list containing the interval data

distribution
a string vector indicating the distributions with which to fit the data

stan_opts
a named list with additional options passed to CmdStanR

return_raw
a logical, should the samples be returned